changed up the build process

+++ tests/jmh/native-related-runtime/buildrun.sh
@@ -2,6 +2,9 @@
 
 #ignoreLock=-Djmh.ignoreLock=true
 
+mvn clean install && ./transferjars.sh ../../../src/java/target/jRAPL-1.0.jar target/benchmarks.jar
+[ $? == 0 ] || exit
+
 rm -rf data && mkdir data # where results will be written to
 sudo java -jar $ignoreLock target/benchmarks.jar
